Move bool_expression's destructor fully inline
This is necessary to allow the compiler to completely eliminate calls to
the destructor. For Clang this significantly improves the code locality
of the non-failure path. Before this change there's several branches
that get taken when not failing. With this change there's only one
branch instruction left that gets taken only when failing.