Split policy (to use colors) from mechanism (is it a TTY)
First add an extension point AwesomeAssert::use_colors(std::ostream&)
that can be overridden by the user to determine whether AwesomeAssert
should use colors when writing to the given stream.

Secondly make the mechanism to determine the file descriptor associated
with a given stream available externally.

The combination of these two should allow the user to extend, in
addition to replacing, the policy decision of whether to use colors.