feat: support a single logical operator
Unfortunately expression decomposition won't work with multiple logical
operators (due to operator precedence), so this is the best we can do:

    ASSERT(a <= x && x <= b);
    ASSERT(y <  b || b <  y);

But not:

    ASSERT(a == 0 || a <= x && x <= b);

This would cause the center `a <= x` comparison expression to be
collapsed to a prvalue expression of type `bool`. So we currently detect
this by treating an *exact* type of `bool`, i.e. *not* a reference to
`bool`, as the type to the logical operators and produce a readable
compile-time error message in that case (static_assert).